+++ login-workflow/src/screens/ContactScreen/ContactSupportScreenBase.tsx
@@ -12,27 +12,6 @@ import Box, { BoxProps } from '@mui/material/Box';
 import { ContactScreenClassKey, getContactScreenUtilityClass } from './utilityClasses';
 import { unstable_composeClasses as composeClasses } from '@mui/base';
 
-/**
- * Component renders a screen with contact information for support with the application.
- * Contact information is pulled from the context passed into the workflow.
- *
- * @param icon the icon to display in the header
- * @param emailSupportTitle text to display as the contact support title
- * @param emailSupportContent content to display as the contact support content
- * @param phoneSupportTitle text to display as the phone support title
- * @param phoneSupportContent content to display as the phone support content
- * @param contactEmail to display the contact email
- * @param contactPhone to display contact phone number
- * @param dismissButtonLabel to display label for the button
- * @param onDismiss function to call when user clicks button
- * @param WorkflowCardBaseProps props that will be passed to the WorkflowCard component
- * @param WorkflowCardHeaderProps props that will be passed to the WorkflowCardHeader component
- * @param WorkflowCardInstructionProps props that will be passed to the WorkflowCardInstructions component
- * @param WorkflowCardActionsProps props that will be passed to the WorkflowCardActions component
- *
- * @category Component
- */
-
 const useUtilityClasses = (ownerState: ContactSupportScreenProps & BoxProps): Record<ContactScreenClassKey, string> => {
     const { classes } = ownerState;
 
@@ -51,6 +30,15 @@ const useUtilityClasses = (ownerState: ContactSupportScreenProps & BoxProps): Re
 
     return composeClasses(slots, getContactScreenUtilityClass, classes);
 };
+
+/**
+ * Component renders a screen with contact information for support with the application.
+ * Contact information is pulled from the context passed into the workflow.
+ *
+ * @param {ContactSupportScreenProps} props - props of contactSupportScreen
+ *
+ * @category Component
+ */
 
 export const ContactSupportScreenBase: React.FC<ContactSupportScreenProps> = (props) => {
     const {
