+++ login-workflow/src/screens/CreatePasswordScreen/CreatePasswordScreen.tsx
@@ -9,12 +9,7 @@ import { useTranslation } from 'react-i18next';
 /**
  * The component renders a screen with the password and confirm password field for creating a new password.
  *
- * @param errorDisplayConfig configuration for customizing how errors are displayed
- * @param PasswordProps props passed from SetPassword component
- * @param WorkflowCardBaseProps props that will be passed to the WorkflowCard component
- * @param WorkflowCardHeaderProps props that will be passed to the WorkflowCardHeader component
- * @param WorkflowCardInstructionProps props that will be passed to the WorkflowCardInstructions component
- * @param WorkflowCardActionsProps props that will be passed to the WorkflowCardActions component
+ * @param {CreatePasswordScreenProps} props - props of createPasswordScreen
  *
  * @category Component
  */
