+++ login-workflow/src/screens/EulaScreen/EulaScreen.tsx
@@ -8,18 +8,7 @@ import { useTranslation } from 'react-i18next';
 /**
  * Component that renders a screen displaying the EULA and requests acceptance via a checkbox.
  *
- * @param eulaContent the content to render for the EULA. Can be a plain string or HTML
- * @param html true if the EULA should be rendered as HTML
- * @param checkboxLabel label for the EULA checkbox
- * @param initialCheckboxValue used to pre-populate the checked/unchecked checkbox when the screen loads
- * @param checkboxProps used to set checkbox props
- * @param onEulaAcceptedChange used to test eula checkbox accepted
- * @param WorkflowCardBaseProps props that will be passed to the WorkflowCard component
- * @param WorkflowCardHeaderProps props that will be passed to the WorkflowCardHeader component
- * @param WorkflowCardInstructionProps props that will be passed to the WorkflowCardInstructions component
- * @param WorkflowCardActionsProps props that will be passed to the WorkflowCardActions component
- * @param errorDisplayConfig configuration for customizing how errors are displayed
- * @param refreshConfig used to refresh Eula content.
+ * @param {EulaScreenProps} props - props of eulaScreen
  *
  * @category Component
  */
